// src/app/services/api/interfaces/contas/CS_IReqSave1202.ts
export interface CS_IReqSave1202 {
    Id: string
    BB012_CNPJ: number
    BB012_InscEstadual: number
    BB012_SUFRAMA: string
    BB012_RegSUFRAMAValido: number
    BB012_RegJuntaComercial: string
    BB012_DataRegJunta: string
    BB012_Patrimonio: number
    BB012_Capital_Social: number
    BB012_CPF: number
    BB012_RG: number
    BB012_ComplementoRG: string
    BB012_EmissaoRG: string
    BB012_PIS: number
    BB012_ResideDesde: string
    BB012_NroDependentes: number
    BB012_EmpAdmissao: string
    BB012_Emp_Profissao: string
    BB012_ValorRemuneracao: number
    BB012_OutrosRendimentos: number
    BB012_OrigemOutrosRend: string
    BB012_Insc_Est_SNI_ID: number
    BB012_Sexo_ID: number
    BB012_EstadoCivil_ID: number
    BB012_TipoDomicilio_ID: number
    BB012_CompResid01_ID: number
    BB012_CompResid02_ID: number
    BB012_GEscolaridade_ID: number
    BB012_Ocupacao_Id: number
    BB012_NaturalDe_ID: string
    BB012_TpTributacao_ID: number
    BB012_Ident_Estrangeiro: string
    BB012_Empresa: string
    BB012_Emp_Endereco: string
    BB012_Emp_Grupo_ID: number
    BB012_MotDesoneracaoID: number
}