// src/app/services/api/interfaces/contas/CS_IReqSaveConta.ts
export interface IReqSaveConta {
    ID?: string
    BB012_Codigo?: number
    BB012_Nome_Cliente?: string
    BB012_Nome_Fantasia?: string
    BB012_Data_Aniversario?: string
    BB012_Data_Cadastro?: string
    BB012_Telefone?: string
    BB012_FaxCelular?: string
    BB012_Home_Page?: string
    BB012_Email?: string
    BB012_Data_Entrada_Sit?: string
    BB012_Data_Saida_Sit?: string
    BB012_Descricao?: string
    BB012_Is_Active?: boolean
    BB012_Tipo_Conta_ID?: number
    BB012_Grupoconta_ID?: number
    BB012_ClasseConta_ID?: number
    BB012_StatusConta_ID?: number
    BB012_Sit_Conta_ID?: number
    BB012_ModRelacao_ID?: number
    BB012_Sequence?: number
    bb012_dUltAlteracao?: string
    bb012_EstabCadID?: string
    bb012_KeyAcess?: string
    bb012_ID_Indicador?: string
    bb012_CountAppMCon?: number
    bb012_OriCadastroID?: number
}